% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selk.R
\name{piv_k}
\alias{piv_k}
\title{Selection of the number of mixture components k via
leave-one-out (loo) cross-validation.}
\usage{
piv_k(y, priors, ic = c("dic", "loo"), nMC = 500, chains = 4,
  k_max = 8, burn = 0.5 * nMC)
}
\description{
Selection of the number of mixture components k via
leave-one-out (loo) cross-validation.
}
