% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{piv_plot}
\alias{piv_plot}
\title{Plotting outputs from pivotal relabelling}
\usage{
piv_plot(y, res, rel, type)
}
\arguments{
\item{y}{Data vector or matrix.}

\item{type}{Type of plots required. Choose among: \code{"chains"}, \code{"estimates"}, \code{"estimates_hist"}.}

\item{est}{Pivotal estimates as provided by \code{piv_rel}.}

\item{n.iter}{Number of valid MCMC iterations as provided by \code{piv rel}.}

\item{switch.means}{Post-processed chains.}

\item{true.means}{An estimate for the true means.}
}
\description{
Plot and visualize MCMC outputs, posterior relabelled chains and estimates and diagnostics.
}
\examples{

Fishery data

data(fish)
y <- fish[,1]
N <- length(y)
k <- 5
nMC <- 5000
res <- piv_MCMC(y, k, nMC)
rel <- piv_rel(res$mu_switch, res$groupPost, res$clust_sel,
                               Mu=res$Mu,
                               nMC = nMC)
piv_plot(y, res, rel, "chains")

piv_plot(y, res, rel, "estimates")

piv_plot(y, res, rel, "estimates_hist")


}
