% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relab.R
\name{piv_rel}
\alias{piv_rel}
\title{Perfroming the pivotal relabelling step and computing the relabelled posterior estimates}
\usage{
piv_rel(mcmc, nMC)
}
\arguments{
\item{mcmc}{The output of the MCMC sampling from \code{piv_MCMC}.}

\item{nMC}{The number of total MCMC iterations (given in input to the \code{piv_MCMC} function, or any function suited for MCMC sampling).}
}
\value{
This function gives the relabelled posterior estimates--both mean and medians--obtained from the Markov chains of the MCMC sampling.

\item{\code{mu_rel_mean}}{ Estimated posterior means}
\item{\code{mu_rel_median}}{ Estimated posterior medians}
\item{\code{mu_rel_complete}}{Complete relabelled chains}
\item{\code{Final_It}}{The final number of valid iterations}
}
\description{
This function allows to perform the pivotal relabelling procedure described in Egidi et al. (2018) and to obtain the relabelled posterior estimates.
}
\details{
Prototypical models in which the label switching problem arises are mixture models, where for a sample \eqn{y=(y_1,\ldots,y_n)} we assume
 \deqn{
(Y_i|Z_i=j) \sim f(y;\mu_j,\phi),
}
where the \eqn{Z_i}, \eqn{i=1,\ldots,n}, are i.i.d. random variables, \eqn{j=1,\dots,k},
\eqn{\phi} is a parameter which is common to all components,  \eqn{Z_i \in {1,\ldots,k}},
and
\deqn{
P(Z_i=k)=\pi_k.
}

This model is unidentified with respect to an arbitrary permutation of the labels \eqn{1,...,k}. Relabelling means permuting
the labels at each iteration of the Markov chain in such
a way that the relabelled chain can be used to draw inferences
on component-specific parameters.


We assume here that an MCMC sample is obtained from the posterior distribution for model above--for instance via \code{piv_MCMC} function--with a prior distribution which is labelling invariant.
Furthermore, suppose that we can find \eqn{k} units, one
for each group, which are (pairwise) separated with (posterior) probability one
(that is, the posterior probability of any two of them being in the same group
is zero).
It is then straightforward to use the \eqn{k} units, called pivots in what follows, to identify the groups and to relabel the chains (see the vignette for thorough details).
}
\examples{

#Univariate simulation

N   <- 250
nMC <- 2500
k   <- 3
p   <- rep(1/k,k)
x   <- 3
stdev <- cbind(rep(1,k), rep(200,k))
Mu    <- seq(-trunc(k/2)*x,trunc(k/2)*x,length=k)
W     <- c(0.2,0.8)
sim   <- piv_sim(N,k,Mu,stdev,W=W)
res   <- piv_MCMC(sim$y, k, nMC)
rel   <- piv_rel(mcmc=res, nMC = nMC)


#Bivariate simulation

N <- 200
k <- 3
nMC <- 5000
M1  <- c(-.5,8)
M2  <- c(25.5,.1)
M3  <- c(49.5,8)
Mu  <- matrix(rbind(M1,M2,M3),c(k,2))
stdev <- cbind(rep(1,k), rep(200,k))
Sigma.p1 <- matrix(c(stdev[1,1],0,0,stdev[1,1]),
                   nrow=2, ncol=2)
Sigma.p2 <- matrix(c(stdev[1,2],0,0,stdev[1,2]),
                   nrow=2, ncol=2)
W <- c(0.2,0.8)
sim <- piv_sim(N,k,Mu,stdev,Sigma.p1,Sigma.p2,W)
res <- piv_MCMC(sim$y, k, nMC)
rel <- piv_rel(mcmc = res, nMC = nMC)
piv_plot(y=sim$y, mcmc=res, rel_est = rel, type="chains")
piv_plot(y=sim$y, mcmc=res, rel_est = rel,
         type="estimates_hist")


}
\references{
Egidi, L., Pappada, R., Pauli, F. and Torelli, N. (2018). Relabelling in Bayesian Mixture
Models by Pivotal Units. Statistics and Computing, 28(4), 957-969, DOI 10.1007/s11222-017-  9774-2.
}
\author{
Leonardo Egidi \url{legidi@units.it}
}
