% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotal.R
\name{piv_sel}
\alias{piv_sel}
\title{Pivotal Selection via Co-Association Matrix}
\usage{
piv_sel(k, gIndex, C, N, ZM)
}
\arguments{
\item{k}{The number of mixture components/groups.}

\item{gIndex}{Clusters' allocation.}

\item{C}{Co-association matrix.}

\item{N}{Data sample size}

\item{ZM}{Auxiliary matrix used for building \code{C}.}
}
\value{
\item{\code{pivots}}{ The pivotal units. }
}
\description{
Finding the pivots according to three different
methods involving a co-association matrix C.
}
\details{
Given \eqn{H} distinct partitions of \eqn{N} units in \eqn{k} groups,
we can build a co-association matrix \eqn{C},
where \eqn{c_{i,p}=n_{i,p}/H}, with \eqn{n_{i,p}} the number of times
the pair \eqn{(y_{i},y_{p})} is assigned to the same
cluster among the \eqn{H} partitions.

Let \eqn{j} be the group containing units \eqn{\mathcal J_j},
the user may choose \eqn{{i^*}\in\mathcal J_j} that
maximizes one of the quantities:
\deqn{
 \sum_{p\in\mathcal J_j} c_{{i^*}p}}

 or
 \deqn{\sum_{p\in\mathcal J_j} c_{{i^*}p} - \sum_{j\not\in\mathcal J_j} c_{{i^*}p}.
}

These methods give the unit that maximizes the global
within similarity (\code{"maxsumint"} and the unit that
maximizes the difference between global within and
between similarities \code{"maxsumdiff"}, respectively.
Alternatively, we may choose \eqn{i^{*} \in\mathcal J_j}, which minimizes:
\deqn{\sum_{p\not\in\mathcal J_j} c_{i^{*}p},}
obtaining the most distant unit among the members
that minimize the global dissimilarity between one group
and all the others (\code{"maxsumnoint"}).
}
