% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{piv_sim}
\alias{piv_sim}
\title{Generate Data from a Gaussian Nested Mixture}
\usage{
piv_sim(N, k, Mu, stdev, Sigma.p1 = matrix(c(1, 0, 0, 1), 2, 2, byrow =
  TRUE), Sigma.p2 = matrix(c(100, 0, 0, 100), 2, 2, byrow = TRUE),
  W = c(0.5, 0.5))
}
\arguments{
\item{N}{Sample size, data dimension.}

\item{k}{Number of mixture components.}

\item{Mu}{Initial mean vector/matrix.}

\item{stdev}{Initial standard deviations (for univariate mixtures).}

\item{Sigma.p1}{Covariance matrix for the first mixture level (for bivariate mixtures only).}

\item{Sigma.p2}{Covariance matrix for the second mixture level (for bivariate mixture only).}

\item{W}{Mixture weights for the two levels,vector.}
}
\value{
\item{ \code{y}}{Data values.}
}
\description{
Simulate N observations from a nested Gaussian mixture model with k pre-specified components.
}
\details{
The functions allows to simulate values from a double (nested) univariate
Gaussian mixture:

\deqn{
(Y_i|Z_i=j) sim \sum_{s=1}^{2} p_{js}\, \mathcal{N}(\mu_{j}, \sigma^{2}_{js}),
}

or from a bivariate nested Gaussian mixture;

\deqn{
(Y_i|Z_i=j) sim \sum_{s=1}^{2} p_{js}\, \mathcal{N}_{2}(\bm{\mu}_{j}, \Sigma_{s}).
}

\code{Mu} is the mean input vector/matrix, \code{stdev} is a \eqn{k x 2}
matrix for the standard deviations of the \eqn{k} groups. \code{Sigma.p1} and
\code{Sigma.p2} are the covariances matrices for subgroups 1 and 2,
respectively. \code{W} is a vector of dimension 2 for the subgroups weights.
}
\examples{

# Bivariate mixture simulation with three components

N  <- 2000
k  <- 3
M1 <- c(-45,8)
M2 <- c(45,.1)
M3 <- c(100,8)
Mu <- matrix(rbind(M1,M2,M3),c(k,2))
stdev    <- cbind(rep(1,k), rep(200,k))
Sigma.p1 <- matrix(c(stdev[1,1],0,0,stdev[1,1]),
nrow=2, ncol=2)
Sigma.p2 <- matrix(c(stdev[1,2],0,0,stdev[1,2]),
 nrow=2, ncol=2)
W   <- c(0.2,0.8)
sim <- piv_sim(N,k,Mu,stdev,Sigma.p1,Sigma.p2,W)
plot(sim$y, xlab="y[,1]", ylab="y[,2]")

}
