% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotal.R
\name{pivotal_selection}
\alias{pivotal_selection}
\title{Pivotal Selection via Co-Association Matrix}
\usage{
pivotal_selection(Obj, k, gIndex, C, n, ZM, maxima, available_met)
}
\arguments{
\item{Obj}{Numerical string for the allowed pivotal criterion.}

\item{k}{The number of mixture components/groups.}

\item{gIndex}{Clusters' allocation.}

\item{C}{Co-association matrix.}

\item{n}{Data samle size}

\item{ZM}{Auxiliary matrix used for building \code{C}.}

\item{maxima}{Initial assignment for MUS algorithm.}

\item{available_met}{Available criteria methods (integer).}
}
\value{
\item{\code{Cg}}{ The pivotal units. }
}
\description{
Pivotal selection function - Find the pivots according to four different
methods involving a co-association matrix C.
}
