% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pivotal}
\alias{plot_pivotal}
\title{Plotting outputs from pivotal relabelling}
\usage{
plot_pivotal(y, est, chains, type, mu_switch, n.iter, true.means)
}
\arguments{
\item{y}{Data vector or matrix.}

\item{est}{Pivotal estimates as provided by \code{pivotal_relabelling}.}

\item{type}{Type of plots required. Choose among: \code{"chains"}, \code{"estimates"}.}

\item{n.iter}{Number of valid MCMC iterations as provided by \code{pivotal relabelling}.}

\item{true.means}{An estimate for the true means.}

\item{switch.means}{Post-processed chains.}
}
\description{
Plot and visualize MCMC outputs, posterior relabelled chains and estimates and diagnostics.
}
\examples{

Fishery data

library(bayesmix)
data(fish)
y <- fish[,1]
k <- 5
nMC <- 5000
output_bayes <- bayesMCMC(y, k, nMC)
relab_est <- pivotal_relabelling(output_bayes$mu_switch,
                               output_bayes$groupPost,
                               output_bayes$clust_sel,
                               Mu=output_bayes$Mu,
                               nMC = nMC)

plot_pivotal(y, relab_est$mu_rel_median,
            relab_est$mu_rel_complete,
            type="chains",
            output_bayes$mu_switch,
            relab_est$Final_it, output_bayes$Mu)

plot_pivotal(y, relab_est$mu_rel_median,
            relab_est$mu_rel_complete,
            type="estimates",
            output_bayes$mu_switch,
            relab_est$Final_it, output_bayes$Mu)

plot_pivotal(y, relab_est$mu_rel_median,
            relab_est$mu_rel_complete,
            type="estimates_hist",
            output_bayes$mu_switch,
            relab_est$Final_it, output_bayes$Mu)


}
