% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mus.R
\name{MUS}
\alias{MUS}
\title{MUS algorithm}
\usage{
MUS(C, clusters, prec_par)
}
\arguments{
\item{C}{Square symmetrix matrix with value bounded in \code{[0,1]}. For instance, a co-association matrix resulting from clustering ensembles.}

\item{clusters}{An initial group assignment for the \code{N} statistical units in \code{k} groups.}

\item{prec_par}{A precision parameter for exploring a greater number of algorithm solutions. Default value is 5.}
}
\value{
\item{\code{maxima}}{ The \code{k} maxima units}
}
\description{
Finding the pivotal units through a sequential search in the symmetric matrix C
}
\details{
See the vignette.
}
\examples{
N <- 20
H <- 1000
a <- matrix(NA, H, N)

for (h in 1:H){
   a[h,] <- kmeans(x,centers)$cluster
}
# build the similarity matrix
sim_matr <- matrix(1, N,N)
for (i in 1:(N-1)){
  for (j in (i+1):N){
     sim_matr[i,j] <- sum(a[,i]==a[,j])/H
     sim_matr[j,i] <- sim_matr[i,j]
     }
}

cl <- KMeans(x, centers)$cluster
mus_alg <- MUS(C = sim_matr, clusters = cl, prec_par = 5)


}
