% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{piv_plot}
\alias{piv_plot}
\title{Plotting outputs from pivotal relabelling}
\usage{
piv_plot(y, est, chains, type, mu_switch, n.iter, true.means)
}
\arguments{
\item{y}{Data vector or matrix.}

\item{est}{Pivotal estimates as provided by \code{pivotal_relabelling}.}

\item{type}{Type of plots required. Choose among: \code{"chains"}, \code{"estimates"}.}

\item{n.iter}{Number of valid MCMC iterations as provided by \code{pivotal relabelling}.}

\item{true.means}{An estimate for the true means.}

\item{switch.means}{Post-processed chains.}
}
\description{
Plot and visualize MCMC outputs, posterior relabelled chains and estimates and diagnostics.
}
\examples{

Fishery data

library(bayesmix)
data(fish)
y <- fish[,1]
N <- length(y)
k <- 5
nMC <- 5000
res <- piv_MCMC(y, k, nMC)
rel <- piv_rel(res$mu_switch, res$groupPost, res$clust_sel,
                               Mu=res$Mu,
                               nMC = nMC)

piv_plot(y, rel$mu_rel_median, rel$mu_rel_complete,
type="chains", res$mu_switch, rel$Final_it, res$Mu)

piv_plot(y, rel$mu_rel_median, rel$mu_rel_complete,
type="estimates", res$mu_switch, rel$Final_it, res$Mu)

piv_plot(y, rel$mu_rel_median, rel$mu_rel_complete,
type="estimates_hist", res$mu_switch, rel$Final_it, res$Mu)


}
