% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{piv_plot}
\alias{piv_plot}
\title{Plotting outputs from pivotal relabelling}
\usage{
piv_plot(y, mcmc, rel_est, type = c("chains", "estimates", "hist"))
}
\arguments{
\item{y}{Data vector or matrix.}

\item{rel_est}{Pivotal estimates as provided by \code{piv_rel}.}

\item{type}{Type of plots required. Choose among: \code{"chains"}, \code{"estimates"}, \code{"hist"}.}
}
\description{
Plot and visualize MCMC outputs, posterior relabelled chains and estimates and diagnostics.
}
\examples{

# Fishery data

data(fish)
y <- fish[,1]
N <- length(y)
k <- 5
nMC <- 5000
res <- piv_MCMC(y = y, k = k, nMC = nMC)
rel <- piv_rel(mcmc=res, nMC = nMC)
piv_plot(y, res, rel, "chains")
piv_plot(y, res, rel, "estimates")
piv_plot(y, res, rel, "hist")

}
\author{
Leonardo Egidi \url{legidi@units.it}
}
