% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{piv_plot}
\alias{piv_plot}
\title{Plotting outputs from pivotal relabelling}
\usage{
piv_plot(y, mcmc, rel_est, type = c("chains", "estimates", "hist"))
}
\arguments{
\item{y}{Data vector or matrix.}

\item{mcmc}{The ouptut of the raw MCMC sampling, as provided by \code{piv_MCMC}.}

\item{rel_est}{Pivotal estimates as provided by \code{piv_rel}.}

\item{type}{Type of plots required. Choose among: \code{"chains"}, \code{"estimates"}, \code{"hist"}.}
}
\description{
Plot and visualize MCMC outputs, posterior relabelled chains and estimates and diagnostics.
}
\author{
Leonardo Egidi \url{legidi@units.it}
}
