% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotal.R
\name{piv_sel}
\alias{piv_sel}
\title{Pivotal Selection via Co-Association Matrix}
\usage{
piv_sel(k, gIndex, C, n, ZM)
}
\arguments{
\item{k}{The number of mixture components/groups.}

\item{gIndex}{Clusters' allocation.}

\item{C}{Co-association matrix.}

\item{n}{Data sample size}

\item{ZM}{Auxiliary matrix used for building \code{C}.}
}
\value{
\item{\code{pivots}}{ The pivotal units. }
}
\description{
Finding the pivots according to four different
methods involving a co-association matrix C. This is an internal function launched by \code{piv_MCMC}.
}
\details{
Let \eqn{j} be the group containing units \eqn{\mathcal J_j},
the user may chose \eqn{{i^*}\in\mathcal J_j} that
maximizes one of the quantities:
\deqn{
 \sum_{p\in\mathcal J_j} c_{{i^*}p}
 \sum_{p\in\mathcal J_j} c_{{i^*}p} - \sum_{j\not\in\mathcal J_j} c_{{i^*}p}.
}

These methods give the unit that maximizes the global
within similarity (\code{"maxsumint"} and the unit that
maximizes the difference between global within and
between similarities \code{"maxsumdiff"}, respectively.
Alternatively, we may choose \eqn{i^{*} \in\mathcal J_j}, which minimizes:
\deqn{\sum_{p\not\in\mathcal J_j} c_{i^{*}p},}
obtaining the most distant unit among the members
that minimize the global dissimilarity between one group
and all the others (\code{"maxsumnoint"}).
}
