% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotal.R
\name{piv_sel}
\alias{piv_sel}
\title{Pivotal Selection via Co-Association Matrix}
\usage{
piv_sel(k, gIndex, C, n, ZM)
}
\arguments{
\item{k}{The number of mixture components/groups.}

\item{gIndex}{Clusters' allocation.}

\item{C}{Co-association matrix.}

\item{n}{Data sample size}

\item{ZM}{Auxiliary matrix used for building \code{C}.}
}
\value{
\item{\code{pivots}}{ The pivotal units. }
}
\description{
Finding the pivots according to four different
methods involving a co-association matrix C. This is an internal function launched by \code{piv_MCMC}.
}
