% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{piv_sim}
\alias{piv_sim}
\title{Generate Data from a Gaussian Nested Mixture}
\usage{
piv_sim(N, k, Mu, stdev, Sigma.p1, Sigma.p2, W)
}
\arguments{
\item{N}{Sample size, data dimension.}

\item{k}{Number of mixture components.}

\item{Mu}{Initial mean vector/matrix.}

\item{stdev}{Initial standard deviations (for univariate mixtures).}

\item{Sigma.p1}{Covariance matrix for the first mixture level (for bivariate mixtures only).}

\item{Sigma.p2}{Covariance matrix for the second mixture level (for bivariate mixture only).}

\item{W}{Mixture weights for the two levels,vector.}
}
\value{
\code{y}  Data values.
}
\description{
Simulate N observations from a nested Gaussian mixture model with k pre-specified components.
}
\examples{

Bivariate mixture simulation with three components

N  <- 200
k  <- 3
M1 <- c(-.5,8)
M2 <- c(25.5,.1)
M3 <- c(49.5,8)
Mu <- matrix(rbind(M1,M2,M3),c(k,2))
stdev    <- cbind(rep(1,k), rep(200,k))
Sigma.p1 <- matrix(c(stdev[1,1],0,0,stdev[1,1]),
nrow=2, ncol=2)
Sigma.p2 <- matrix(c(stdev[1,2],0,0,stdev[1,2]),
 nrow=2, ncol=2)
W   <- c(0.2,0.8)
sim <- piv_sim(N,k,Mu,stdev,Sigma.p1,Sigma.p2,W)

}
